Consolidate distutils importing hacks into _distutils_importer package. Generate distutils-precedence.pth inline.
