+++ MANIFEST.in
@@ -13,5 +13,4 @@ include launcher.c
 include msvc-build-launcher.cmd
 include pytest.ini
 include tox.ini
-include distutils_precedence.pth
 exclude pyproject.toml  # Temporary workaround for #1644.
