+++ _distutils_importer/install.py
@@ -1,0 +1,5 @@
+from . import enabled, add_shim
+
+
+if enabled():
+    add_shim()
